deepclean_requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==24.2.0
blobfile==3.0.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
certifi==2024.6.2
cffi @ file:///croot/cffi_1726856441404/work
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
conda @ file:///croot/conda_1729193882500/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1724940562255/work
cycler==0.12.1
DataProperty==1.1.0
datasets==3.1.0
defusedxml==0.7.1
diffusers==0.29.0
dill==0.3.8
distro @ file:///croot/distro_1714488253808/work
docker-pycreds==0.4.0
durationpy==0.9
evaluate==0.4.3
fastapi==0.115.12
fastargs @ git+https://github.com/jinghanjia/fastargs@bde495ba4bed446b7695d1347953ee005a566a2e
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
fonttools==4.57.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.39.0
gradio==5.0.2
gradio_client==1.4.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==7.1.0
iopath==0.1.10
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.2
jsonlines==4.0.0
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
-e git+https://github.com/luca-medeiros/lang-segment-anything@3a3868bbf01774064f9a5b10709aee297c14db1c#egg=lang_sam
libmambapy @ file:///croot/mamba-split_1725563367192/work/libmambapy
litserve==0.2.3
llama_stack==0.2.5
llama_stack_client==0.2.5
-e git+https://github.com/OPTML-Group/SOUL.git@30f7a416d021c7f63ec0fe33569b3d7decefd804#egg=lm_eval&subdirectory=lm-evaluation-harness
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mbstrdecoder==1.1.4
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1723567589013/work
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.77.0
opencv-python-headless==4.10.0.84
orjson==3.10.18
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.2
pathvalidate==3.2.3
peft==0.15.2
pillow==10.2.0
platformdirs==4.2.2
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
portalocker==3.1.1
prompt_toolkit==3.0.51
propcache==0.2.1
protobuf==5.27.1
psutil==6.0.0
pyaml==25.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
regex==2024.5.15
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
ruff==0.11.9
sacrebleu==2.5.1
safetensors==0.4.3
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2@c2ec8e14a185632b0a5d8b161928ceb50197eddc
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
sentry-sdk==2.6.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.46.2
supervision==0.23.0
sympy==1.12
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.4
tqdm-multiprocess==0.0.11
transformers==4.44.2
triton==3.0.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typepy==1.3.4
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.2
uvloop==0.20.0
wandb==0.17.2
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wheel==0.44.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.19.2
zstandard @ file:///croot/zstandard_1728569189425/work
